Завершено
Закончил создание точек вызова исключений. Все они находятся внутри соответствующих моделей.  это помогает не усложнять понимание кода и не даст выйти исключениям на уровень вывыше...
Достоинство этого реешения:
- не усложняется код контроллеров
- позволяет просто и легко изменять код

+++ MishpahugHaifa/src/main/java/application/models/kichentype/KichenTypeModel.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import javax.transaction.Transactional;
 
+import application.exceptions.EntityExistsException;
+import application.exceptions.NotFoundEntityException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,11 +21,14 @@ public class KichenTypeModel implements IKichenTypeModel {
 
     @Override
     public KitchenTypeEntity getById(Integer id) {
+        if (!kichenTypeRepository.existsById(id)) throw new NotFoundEntityException("");
         return kichenTypeRepository.getOne(id);
     }
 
     @Override
     public KitchenTypeEntity updateName(Integer id, String name) {
+        if (!kichenTypeRepository.existsById(id)) throw new NotFoundEntityException("");
+        if (kichenTypeRepository.existsByName(name)) throw new EntityExistsException("");
         KitchenTypeEntity countryEntity = kichenTypeRepository.getOne(id);
         countryEntity.setName(name);
         return kichenTypeRepository.save(countryEntity);
@@ -31,6 +36,7 @@ public class KichenTypeModel implements IKichenTypeModel {
 
     @Override
     public void deleteByName(String name) {
+        if (!kichenTypeRepository.existsByName(name)) throw new NotFoundEntityException("");
         kichenTypeRepository.deleteByName(name);
     }
 
@@ -46,11 +52,14 @@ public class KichenTypeModel implements IKichenTypeModel {
 
     @Override
     public KitchenTypeEntity add(KitchenTypeEntity data) {
+        if (kichenTypeRepository.existsByName(data.getName())) throw new EntityExistsException("");
         return kichenTypeRepository.save(data);
     }
 
     @Override
     public KitchenTypeEntity update(Integer id, String name) {
+        if (!kichenTypeRepository.existsById(id)) throw new NotFoundEntityException("");
+        if (kichenTypeRepository.existsByName(name)) throw new EntityExistsException("");
         KitchenTypeEntity temp = kichenTypeRepository.getOne(id);
         temp.setName(name);
         return kichenTypeRepository.save(temp);
@@ -58,6 +67,7 @@ public class KichenTypeModel implements IKichenTypeModel {
 
     @Override
     public KitchenTypeEntity getByName(String kichenType) {
+        if (!kichenTypeRepository.existsByName(kichenType)) throw new NotFoundEntityException("");
         return kichenTypeRepository.getByName(kichenType);
     }
 }
