+++ MishpahugHaifa/src/main/java/application/models/marriagestatus/MaritalStatusModel.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import javax.transaction.Transactional;
 
+import application.exceptions.EntityExistsException;
+import application.exceptions.NotFoundEntityException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +21,7 @@ public class MaritalStatusModel implements IMaritalStatusModel {
 
     @Override
     public MaritalStatusEntity getByName(String name) {
+        if (!maritalStatusRepository.existsByName(name)) throw new NotFoundEntityException("");
         return maritalStatusRepository.getByName(name);
     }
 
@@ -30,11 +33,13 @@ public class MaritalStatusModel implements IMaritalStatusModel {
 
     @Override
     public MaritalStatusEntity getById(Integer id) {
+        if (!maritalStatusRepository.existsById(id)) throw new NotFoundEntityException("");
         return maritalStatusRepository.getOne(id);
     }
 
     @Override
     public void deleteByName(String name){
+        if (!maritalStatusRepository.existsByName(name)) throw new NotFoundEntityException("");
         maritalStatusRepository.deleteByName(name);
     }
 
@@ -45,6 +50,8 @@ public class MaritalStatusModel implements IMaritalStatusModel {
 
     @Override
     public MaritalStatusEntity updateName(Integer id, String name){
+        if (maritalStatusRepository.existsByName(name)) throw new EntityExistsException("");
+        if (!maritalStatusRepository.existsById(id)) throw new NotFoundEntityException("");
         MaritalStatusEntity maritalStatusEntity = maritalStatusRepository.getOne(id);
         maritalStatusEntity.setName(name);
         return maritalStatusRepository.saveAndFlush(maritalStatusEntity);
@@ -52,6 +59,7 @@ public class MaritalStatusModel implements IMaritalStatusModel {
 
     @Override
     public MaritalStatusEntity add(MaritalStatusEntity data){
+        if (maritalStatusRepository.existsByName(data.getName())) throw new EntityExistsException("");
         return maritalStatusRepository.saveAndFlush(data);
     }
